JPG version
I transferred it all over to vector format (photoshop to illustrator), so whatever size(s) we decide on, i can upscale them without any loss of image quality. the red buttons are on the left are when you select it, i have gray versions for all of the buttons for when they are not pressed. 

i'm going back to see what the common gametype settings are, and i went back and picked out the Lazer Runner buttons we wouldn't be using.

what gametypes do we want to keep? right now i only did the 3, but i can add more. FFA = Free For All, TDM = Team Deathmatch, and Elim = Eliminator. I like the potential in having the numbers update inside the square representing the teams, but i don't like the font, and i might have to do the boxes better. 

i'll definitely have the gametype options tomorrow, based on the gametypes i have so far. basically a smaller menu will be a layer above the gametype options displaying specific buttons. i'll also have labels that will update based on what is picked.